11_handoffs/pyproject.toml
[project]
name = "11-handoffs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "asyncio>=3.4.3",
    "openai-agents>=0.2.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
]
